Improvements to Settings
More consistency with other buttons.
